Add a metadata to be included in the shiny app

Add a metadata to be included in the shiny app.

add_meta(scConf, meta, obj, maxLevels = 50)

Arguments

scConf

shinycell config data.table

meta

metadata to add from the single-cell metadata. Must match one of the following:

Seurat objects: column names in seu@meta.data i.e. colnames(seu@meta.data)

SCE objects: column names in sce@colData i.e. colnames(sce@colData)

h5ad files: column names in h5ad.obs i.e. h5ad.obs.columns.values

loom files: column names in loom/col_attrs i.e. loom/col_attrs.names

obj

input single-cell object for Seurat (v3+) / SingleCellExperiment data or input file path for h5ad / loom files

maxLevels

maximum number of levels allowed for categorical metadata. Metadata with nlevels > maxLevels will throw up an error message

Value

updated shinycell config data.table

Author

John F. Ouyang

Examples

if (FALSE) {
scConf = add_meta(scConf, c("orig.ident"), seu)
}