Make a shiny app

Make a shiny app based on the shinycell config data.table and single-cell data object.

make_app(
  obj,
  scConf,
  gex.assay = NA,
  gex.slot = c("data", "scale.data", "counts"),
  gene.mapping = FALSE,
  shiny.title = "scRNA-seq shiny app",
  shiny.dir = "shinyApp/",
  shiny.prefix = "sc1",
  enableSubset = TRUE,
  defPtSiz = 1.25,
  default.gene1 = NA,
  default.gene2 = NA,
  default.multigene = NA,
  default.dimred = NA,
  tabs = c("civge", "civci", "gevge", "gem", "gec", "vio", "pro", "hea", "about"),
  theme = "flatly",
  font = "Lato",
  ganalytics = NA
)

Arguments

obj

input single-cell object for Seurat (v3+) / SingleCellExperiment data or input file path for h5ad / loom files

scConf

shinycell config data.table

gex.assay

assay in single-cell data object to use for plotting gene expression, which must match one of the following:

Seurat objects: "RNA" or "integrated" assay, default is "RNA"

SCE objects: "logcounts" or "normcounts" or "counts", default is "logcounts"

h5ad files: "X" or any assay in "layers", default is "X"

loom files: "matrix" or any assay in "layers", default is "matrix"

gex.slot

slot in single-cell assay to plot. This is only used for Seurat objects (v3+). Default is to use the "data" slot

gene.mapping

specifies whether to convert human / mouse Ensembl gene IDs (e.g. ENSG000xxx / ENSMUSG000xxx) into "user-friendly" gene symbols. Set this to TRUE if you are using Ensembl gene IDs. Default is FALSE which is not to perform any conversion. Alternatively, users can supply a named vector where names(gene.mapping) correspond to the actual gene identifiers in the gene expression matrix and gene.mapping correspond to new identifiers to map to

shiny.title

title for shiny app

shiny.dir

specify directory to create the shiny app in. Default is to create a new directory named "shinyApp"

shiny.prefix

specify file prefix

enableSubset

specify whether to enable "Toggle to subset cells" functionality in the shiny app. Default is to enable this functionality

defPtSiz

specify default point size for single cells. For example, a smaller size can be used if you have many cells in your dataset

default.gene1

specify primary default gene to show

default.gene2

specify secondary default gene to show

default.multigene

character vector specifying the default genes to show in bubbleplot / heatmap

default.dimred

character vector specifying the two default dimension reductions. Default is to use UMAP if not TSNE embeddings

tabs

Vector of tab names to include

theme

Bootstrap theme

font

Google font for plots

ganalytics

Google analytics tracking ID (e.g. "UA-123456789-0")

Value

directory containing shiny app

Author

John F. Ouyang

Roy Francis