Generate code files required for shiny app (one dataset)

Generate code files required for shiny app containing only one dataset. In particular, two R scripts will be generated, namely server.R and ui.R. If users want to include multiple dataset in one shiny app, please use make_code_multi() instead. Note that both make_file and make_code functions are ran when running the wrapper function make_app.

make_code(
  shiny.title = "App",
  shiny.prefix = "sc1",
  shiny.dir = "app",
  enableSubset = TRUE,
  defPtSiz = 1.25,
  theme = "flatly",
  tabs = c("civge", "civci", "gevge", "gem", "gec", "vio", "pro", "hea", "about"),
  font = "Lato",
  ganalytics = NA
)

Arguments

shiny.title

title for shiny app

shiny.prefix

specify file prefix

shiny.dir

specify directory to create the shiny app in

enableSubset

specify whether to enable "Toggle to subset cells" functionality in the shiny app. Default is to enable this functionality

defPtSiz

specify default point size for single cells. For example, a smaller size can be used if you have many cells in your dataset

theme

Bootsrap theme

tabs

Vector of tab names to include

font

Google font for plots. Defaults to "Lato". Requires package `showtext`.

ganalytics

Google analytics tracking ID (e.g. "UA-123456789-0")

Value

server.R and ui.R required for shiny app

Author

John F. Ouyang

Roy Francis